refactor: restructure leveled map data architecture
- Move LeveledMapData to separate module with impl files
- Split ApplierData impl blocks into dedicated files
- Change immutable_levels.indexes() return type to Vec
- Consolidate mutex access patterns inside LeveledMapDataInner
- Remove redundant impl_leveled_map_data.rs module

This restructuring improves code organization by separating
concerns and makes the codebase more maintainable with
clearer module boundaries and consistent access patterns.
