+++ src/meta/raft-store/src/applier/applier_data/impl_expire_scoped_view_readonly.rs
@@ -1,0 +1,61 @@
+// Copyright 2021 Datafuse Labs
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use std::io;
+use std::ops::RangeBounds;
+
+use futures_util::StreamExt;
+use futures_util::TryStreamExt;
+use map_api::mvcc;
+use map_api::mvcc::ViewReadonly;
+use map_api::IOResultStream;
+use seq_marked::InternalSeq;
+use seq_marked::SeqMarked;
+use state_machine_api::ExpireKey;
+
+use crate::applier::applier_data::ApplierData;
+use crate::leveled_store::types::Key;
+use crate::leveled_store::types::Namespace;
+
+#[async_trait::async_trait]
+impl mvcc::ScopedViewReadonly<ExpireKey, String> for ApplierData {
+    fn base_seq(&self) -> InternalSeq {
+        self.view.base_seq()
+    }
+
+    async fn get(&self, key: ExpireKey) -> Result<SeqMarked<String>, io::Error> {
+        let got = self.view.get(Namespace::Expire, Key::Expire(key)).await?;
+        Ok(got.map(|x| x.into_expire()))
+    }
+
+    async fn range<R>(
+        &self,
+        range: R,
+    ) -> Result<IOResultStream<(ExpireKey, SeqMarked<String>)>, io::Error>
+    where
+        R: RangeBounds<ExpireKey> + Send + Sync + Clone + 'static,
+    {
+        let start = range.start_bound().cloned();
+        let end = range.end_bound().cloned();
+
+        let start = start.map(Key::Expire);
+        let end = end.map(Key::Expire);
+
+        let strm = self.view.range(Namespace::Expire, (start, end)).await?;
+
+        Ok(strm
+            .map_ok(|(k, v)| (k.into_expire(), v.map(|x| x.into_expire())))
+            .boxed())
+    }
+}
