+++ src/meta/raft-store/src/applier/applier_data/impl_user_scoped_view.rs
@@ -1,0 +1,31 @@
+// Copyright 2021 Datafuse Labs
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use map_api::mvcc;
+use seq_marked::SeqMarked;
+use state_machine_api::MetaValue;
+use state_machine_api::UserKey;
+
+use crate::applier::applier_data::ApplierData;
+use crate::leveled_store::types::Key;
+use crate::leveled_store::types::Namespace;
+use crate::leveled_store::types::Value;
+
+#[async_trait::async_trait]
+impl mvcc::ScopedView<UserKey, MetaValue> for ApplierData {
+    fn set(&mut self, key: UserKey, value: Option<MetaValue>) -> SeqMarked<()> {
+        self.view
+            .set(Namespace::User, Key::User(key), value.map(Value::User))
+    }
+}
