+++ src/meta/raft-store/src/leveled_store/leveled_map/leveled_map_data/impl_view_readonly.rs
@@ -12,22 +12,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::collections::BTreeMap;
 use std::io;
-use std::io::Error;
 use std::ops::RangeBounds;
 use std::sync::Arc;
 
 use futures_util::StreamExt;
 use futures_util::TryStreamExt;
-use log::info;
 use map_api::mvcc;
-use map_api::mvcc::Table;
 use map_api::IOResultStream;
 use seq_marked::InternalSeq;
 use seq_marked::SeqMarked;
 
-use crate::leveled_store::leveled_map::LeveledMapData;
+use crate::leveled_store::leveled_map::leveled_map_data::LeveledMapData;
 use crate::leveled_store::types::Key;
 use crate::leveled_store::types::Namespace;
 use crate::leveled_store::types::Value;
@@ -91,52 +87,5 @@ impl mvcc::ViewReadonly<Namespace, Key, Value> for Arc<LeveledMapData> {
                     .boxed())
             }
         }
-    }
-}
-
-#[async_trait::async_trait]
-impl mvcc::Commit<Namespace, Key, Value> for Arc<LeveledMapData> {
-    async fn commit(
-        &mut self,
-        last_seq: InternalSeq,
-        mut changes: BTreeMap<Namespace, Table<Key, Value>>,
-    ) -> Result<(), Error> {
-        info!(
-            "Committing changes to leveled map data: last_seq={}, changes = {:?}",
-            last_seq, changes
-        );
-        let mut writable = self.writable.lock().unwrap();
-
-        // user map
-
-        let user_updates = changes.remove(&Namespace::User);
-
-        if let Some(updates) = user_updates {
-            let it = updates.inner.into_iter().map(|((k, seq_marked), v)| {
-                ((k.into_user(), seq_marked), v.map(|x| x.into_user()))
-            });
-
-            writable.kv.apply_changes(updates.last_seq, it);
-        }
-
-        // expire map
-
-        let expire_updates = changes.remove(&Namespace::Expire);
-
-        if let Some(updates) = expire_updates {
-            let it = updates.inner.into_iter().map(|((k, seq_marked), v)| {
-                ((k.into_expire(), seq_marked), v.map(|x| x.into_expire()))
-            });
-
-            writable.expire.apply_changes(updates.last_seq, it);
-        }
-
-        // seq
-
-        writable.with_sys_data(|sys_data| {
-            sys_data.update_seq(*last_seq);
-        });
-
-        Ok(())
     }
 }
